Update: ボタンの共通部分を別ファイルに切り出し
