+++ app/src/main/java/com/websarva/wings/android/flat/ui/startup/StartupScreen.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.websarva.wings.android.flat.R
+import com.websarva.wings.android.flat.ui.startup.components.LoginAndSignUpButtons
 import com.websarva.wings.android.flat.ui.theme.FLATTheme
 
 @Composable
@@ -36,7 +37,7 @@ fun StartupScreen(
                     thickness = 4.dp,
                     modifier = Modifier.padding(start = 70.dp, end = 70.dp))
                 Spacer(Modifier.weight(1f))
-                Buttons(
+                LoginAndSignUpButtons(
                     onNavigate = onNavigate
                 /*navController*/)
                 Spacer(Modifier.padding(vertical = 16.dp))
@@ -45,62 +46,6 @@ fun StartupScreen(
     }
 }
 
-@Composable
-fun Buttons(
-    onNavigate: (Int) -> Unit
-/*navController: NavController*/) {
-    Row (
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(start = 32.dp, end = 32.dp),
-        horizontalArrangement = Arrangement.SpaceBetween
-    ){
-        LoginButton(onNavigate)
-        SignUpButton(onNavigate)
-    }
-
-}
-
-@Composable
-fun LoginButton(onNavigate: (Int) -> Unit) {
-
-    Button(
-        onClick = {
-                  /*TODO*/
-                    onNavigate(R.id.loginFragment)
-                  },
-        shape = RoundedCornerShape(20.dp),
-//        colors = FLATTheme.colors.primary
-        colors = textButtonColors(
-            backgroundColor = Color.White,
-            contentColor = Color.Black
-        ),
-        modifier = Modifier.size(width = 140.dp, height = 40.dp),
-        elevation = ButtonDefaults.elevation()
-    ) {
-        Text("ログイン")
-    }
-}
-
-
-@Composable
-fun SignUpButton(
-    onNavigate: (Int) -> Unit
-    /*navController: NavController*/) {
-    Button(
-        onClick = { onNavigate(R.id.accountRegistrationFragment) },
-        shape = RoundedCornerShape(20.dp),
-        colors = textButtonColors(
-            backgroundColor = FLATTheme.colors.primary,
-            contentColor = Color.White
-        ),
-        modifier = Modifier.size(width = 140.dp, height = 40.dp),
-        elevation = ButtonDefaults.elevation()
-
-    ){
-        Text("新規登録")
-    }
-}
 
 //@Preview
 //@Composable
